+++ src/simulation/elements/PBCN.cpp
@@ -120,8 +120,8 @@ static int update(UPDATE_FUNC_ARGS)
 							int r = sim->create_part(-1, x + rx, y + ry, PT_PHOT);
 							if (r != -1)
 							{
-								parts[r].vx = rx * 3;
-								parts[r].vy = ry * 3;
+								parts[r].vx = float(rx * 3);
+								parts[r].vy = float(ry * 3);
 								if (r>i)
 								{
 									// Make sure movement doesn't happen until next frame, to avoid gaps in the beams of photons produced
